+++ test/spec/lib.js
@@ -1,12 +1,18 @@
-const { equal, throws } = require('zoroaster/assert')
-const { resolve } = require('path')
+const { equal, throws, assert } = require('zoroaster/assert')
 const mnpIdio = require('mnp-idio')
+const mnpPackage = require('mnp-package')
 const { findStructure } = require('../../src/lib')
 
 const libTestSuite = {
   'finds a default structure'() {
     const res = findStructure(['node', 'mnp', 'test'])
-    equal(res, `${resolve(__dirname, '../../structures/my-new-package')}/`)
+    const expected = `${mnpPackage}/`
+    equal(res, expected)
+  },
+  'throws an error when -s is the last argument'() {
+    assert.throws(() => {
+      findStructure(['node', 'mnp', 'test', '-s'])
+    })
   },
   'finds an idio structure'() {
     const res = findStructure(['node', 'mnp', 'test', '-s', 'idio'])
@@ -20,10 +26,6 @@ const libTestSuite = {
       args: [['node', 'mnp', 'test', '-s', name]],
       message: `Could not require structure "${name}".`,
     })
-  },
-  'returns default structure when -s is last argument'() {
-    const res = findStructure(['node', 'mnp', 'test', '-s'])
-    equal(res, `${resolve(__dirname, '../../structures/my-new-package')}/`)
   },
 }